+++ HISTORY.md
@@ -10,6 +10,7 @@
 
 ### 2021-05-19
 * gui: disabled scene icon context menu for project scene
+* gui: disabled dragging the project scene icon
 
 ### 2021-05-18
 * gui: fixed exporting media only for a single scene